[⚙️ chore] 대기 데이터를 필터링하는 함수 분리 #23
